requirements.txt
flask==2.3.3
flask-socketio==5.3.4
flask-cors==4.0.0
Pillow==10.4.0
pyautogui==0.9.54
psutil==5.9.5
requests==2.31.0
watchdog==3.0.0
python-dotenv==1.0.0